only increment PC if PC is valid
IFUNCs keep a 0x0 PC.  rb_profile_frames does a check to see if it should record
a frame:

    if (cfp->iseq && cfp->pc) {

With this increment of PC, we cause this check to pass.  We then try to
calculate the line number in calc_lineno:

    size_t pos = (size_t)(pc - iseq->body->iseq_encoded);

This produces a wildly incorrect result for many reasons: iseq_encoded isn't
valid because iseq isn't an rb_iseq_t underneath, producing an essentially
random value, and PC is 0x8, so we underflow and eventually cause an overrun in
succ_index_lookup.